+++ web/src/@types/d3-boxplot.d.ts
@@ -1,0 +1,11 @@
+declare module 'd3-boxplot' {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  function boxplot(): any;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  function boxplotStats(arg: number[]): any;
+
+  export {
+    boxplot,
+    boxplotStats,
+  };
+}
