### 2020-06-22 6fc6978
Add tests against Metaboanalyst functions
As VIIME is a successor to Metaboanalyst, it is useful to verify that it
generates the same results when given the same inputs.

Add tests to verify that the normalization steps shared between
Metaboanalyst and VIIME work the same on both.

Two functions do not function identically: Reference sample
normalization and log transformation. Tests are written for these, but
are currently disabled.
