### 2020-06-23 31174d4
Correct test Metaboanalyst inconsistencies
- The reference sample test assumed that VIIME used PCN, which is not
true. The test was not appropriate and has been removed.
- The Metaboanalyst algorithm for logarithmic transformation is slightly
different for small values, and so requires a larger margin of error.
The test code now supports specifying a margin of error on a CSV.
- Use the builting numpy `log10` and `log2` instead of
`log(...)/log(10)` and `log(...)/log(2)`.
