+++ tests/test_metaboanalyst_comparison.py
@@ -66,19 +66,6 @@ def download_csv(client, csv_id):
 
 
 """
-TODO Reference sample normalization works differently in Metaboanalyst
-Metaboanalyst (https://github.com/xia-lab/MetaboAnalystR/blob/master/R/general_norm_utils.R):
-# normalize by a reference sample (probability quotient normalization)
-# ref should be the name of the reference sample
-ProbNorm<-function(x, ref.smpl){
-    x/median(as.numeric(x/ref.smpl), na.rm=T)
-}
-VIIME (normalization.py):
-def reference_sample(table: pd.DataFrame, argument) -> pd.DataFrame:
-    return table.loc[argument].sum() * table.div(table.sum(axis=1), axis=0)
-"""
-
-"""
 TODO Log transformation works differently in Metaboanalyst
 Metaboanalyst (https://github.com/xia-lab/MetaboAnalystR/blob/master/R/general_norm_utils.R):
 # generalize log, tolerant to 0 and negative values
@@ -91,17 +78,17 @@ def log2(table: pd.DataFrame, min=1e-8):
 """
 
 
-@pytest.mark.parametrize('expected_csv_filename,transform_kwargs', [
-    ('metaboanalyst_normalization_sum.csv', {'normalization': 'sum'}),
-    # ('metaboanalyst_normalization_reference_sample_PIF_178.csv', {
-    # 'normalization': 'reference-sample', 'normalization_arg': 'PIF_178'}),
-    ('metaboanalyst_transformation_cube_root.csv', {'transformation': 'cuberoot'}),
-    # ('metaboanalyst_transformation_log.csv', {'transformation': 'log2'}),
-    ('metaboanalyst_scaling_auto.csv', {'scaling': 'auto'}),
-    ('metaboanalyst_scaling_pareto.csv', {'scaling': 'pareto'}),
-    ('metaboanalyst_scaling_range.csv', {'scaling': 'range'}),
+@pytest.mark.parametrize('expected_csv_filename,transform_kwargs,float_diff', [
+    ('metaboanalyst_normalization_sum.csv', {'normalization': 'sum'}, 1e-9),
+    ('metaboanalyst_transformation_cube_root.csv', {'transformation': 'cuberoot'}, 1e-9),
+    # Metaboanalyst has a slightly different log transformation.
+    # It differs meaningfully at low values, so we need a lower precision for comparisons.
+    ('metaboanalyst_transformation_log.csv', {'transformation': 'log2'}, 0.03),
+    ('metaboanalyst_scaling_auto.csv', {'scaling': 'auto'}, 1e-9),
+    ('metaboanalyst_scaling_pareto.csv', {'scaling': 'pareto'}, 1e-9),
+    ('metaboanalyst_scaling_range.csv', {'scaling': 'range'}, 1e-9),
 ])
-def test_transformations(client, expected_csv_filename, transform_kwargs):
+def test_transformations(client, expected_csv_filename, transform_kwargs, float_diff):
     csv_id = upload_and_transform(client,
                                   'metaboanalyst_original.csv',
                                   **transform_kwargs)
@@ -110,6 +97,6 @@ def test_transformations(client, expected_csv_filename, transform_kwargs):
 
     # Load CSV generated by Metaboanalyst
     with open(metaboanalyst_dir_path / expected_csv_filename, 'r', newline='') as expected_file:
-        expected_csv = CSV(expected_file)
+        expected_csv = CSV(expected_file, float_diff)
 
     assert actual_csv == expected_csv
