### 2020-06-23 #2358 @rowlandm
I could do this,  but the turn around time would be end of July.

Regards,

Rowland Mosbergen

Strategy Analyst | Big Data Analytics & Digital Transformation Lead |
Diversity & Inclusion
B of Eng (Aerospace Avionics)
Mob: 0423 030 626
Email: rowland.mosbergen@gmail.com
Web: https://www.linkedin.com/in/rowlandm-gaicd/

El mié., 24 de jun. de 2020 08:02, Lorena Pantano <notifications@github.com>
escribió:

> Hi @sctyner <https://github.com/sctyner>, @rowlandm
> <https://github.com/rowlandm>, @joannawolthuis
> <https://github.com/joannawolthuis>,
>
> based on your experience maybe you be interested in reviewing this tool?
>
