### 2020-06-25 #2358 @joannawolthuis
Hi Lorena! I'm interested, what are the expectations exactly? :)

---------- Forwarded message ---------
Van: Lorena Pantano <notifications@github.com>
Date: wo 24 jun. 2020 om 00:02
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: Viime: Visualization
and Integration of Metabolomics Experiments (#2358)
Cc: joannawolthuis <jcwolthuis93@gmail.com>, Mention <
mention@noreply.github.com>

Hi @sctyner <https://github.com/sctyner>, @rowlandm
<https://github.com/rowlandm>, @joannawolthuis
<https://github.com/joannawolthuis>,

based on your experience maybe you be interested in reviewing this tool?
