### 2020-06-26 #2358 @lpantano
@joannawolthuis, thank you! this are the guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html and we normally expect reviews in two weeks or so. It should be a pretty nice experience :)
