+++ devops/docker/docker-compose.yml
@@ -1,0 +1,20 @@
+version: "3.8"
+services:
+  opencpu:
+    build:
+      context: ../
+      dockerfile: Dockerfile
+
+  backend:
+    build:
+      context: ../../
+      dockerfile: devops/docker/Dockerfile.backend
+    environment:
+      - MODULE_NAME=viime.wsgi
+      - PORT=5000
+    env_file: 
+      - .env_docker
+    ports:
+      - "5000:5000"
+    volumes:
+      - '../../viime:/app/viime'
