### 2020-07-01 #2358 @rowlandm
Sorry Lorena,

I had to decline as my position is not being extended and I have to look
for work add a priority outside of research software.

Regards,

Rowland Mosbergen

Strategy Analyst | Big Data Analytics & Digital Transformation Lead |
Diversity & Inclusion
B of Eng (Aerospace Avionics)
Mob: 0423 030 626
Email: rowland.mosbergen@gmail.com
Web: https://www.linkedin.com/in/rowlandm-gaicd/

El jue., 25 de jun. de 2020 21:45, Lorena Pantano <notifications@github.com>
escribió:

> Thank you @rowlandm <https://github.com/rowlandm>, I may wait a couple of
> days, but if not I will assign you as a reviewer.
>
