### 2020-07-02 570cc2f
Typescript boxplot (#565)
* Try without D3

* Add grouping back in

* Resize, minor fixes

* Add hover text

* Minor bugs with BoxPlot

* Add Lodash types
