### 2020-07-02 a22e03b
Perform data migrations for row/column information
