+++ migrations/versions/de010dc14402_.py
@@ -7,8 +7,15 @@ Create Date: 2020-07-02 14:17:23.420116
 
 """
 from alembic import op
+from marshmallow import fields, validate
 import sqlalchemy as sa
 import sqlalchemy_utils
+from sqlalchemy_utils.types.json import JSONType
+from sqlalchemy_utils.types.uuid import UUIDType
+
+from viime.models import BaseModel, BaseSchema, CSVFile, db, \
+    METADATA_TYPES, TABLE_COLUMN_TYPES, TABLE_ROW_TYPES, \
+    ValidatedMetaboliteTable
 
 
 # revision identifiers, used by Alembic.
@@ -18,12 +25,115 @@ branch_labels = None
 depends_on = None
 
 
-def upgrade():
-    op.add_column('csv_file', sa.Column('column_json', sqlalchemy_utils.types.json.JSONType(), nullable=False))
-    op.add_column('csv_file', sa.Column('row_json', sqlalchemy_utils.types.json.JSONType(), nullable=False))
+class TableColumn(BaseModel):
+    csv_file_id = db.Column(UUIDType(binary=False), db.ForeignKey('csv_file.id'), primary_key=True)
+    column_index = db.Column(db.Integer, primary_key=True)
+    column_type = db.Column(db.String, nullable=False)
+    subtype = db.Column(db.String, nullable=True)
+    meta = db.Column(JSONType, nullable=True)
+
+    csv_file = db.relationship(
+        CSVFile, backref=db.backref('columns', lazy=True, order_by='TableColumn.column_index'))
+
+
+class TableColumnSchema(BaseSchema):
+    __model__ = TableColumn  # type: ignore
+
+    column_index = fields.Int(required=True, validate=validate.Range(min=0))
+    column_type = fields.Str(required=True, validate=validate.OneOf(TABLE_COLUMN_TYPES))
+    subtype = fields.Str(required=False, validate=validate.OneOf(METADATA_TYPES))
+    meta = fields.Dict(missing=dict)
+
+
+class TableRow(BaseModel):
+    csv_file_id = db.Column(
+        UUIDType(binary=False), db.ForeignKey('csv_file.id'), primary_key=True)
+    row_index = db.Column(db.Integer, primary_key=True)
+    row_type = db.Column(db.String, nullable=False)
+    subtype = db.Column(db.String, nullable=True)
+    meta = db.Column(JSONType, nullable=True)
+
+    csv_file = db.relationship(
+        CSVFile, backref=db.backref('rows', lazy=True, order_by='TableRow.row_index'))
+
+
+class TableRowSchema(BaseSchema):
+    __model__ = TableRow  # type: ignore
+
+    row_name = fields.Str(dump_only=True)
+    row_index = fields.Int(required=True, validate=validate.Range(min=0))
+    row_type = fields.Str(required=True, validate=validate.OneOf(TABLE_ROW_TYPES))
+    subtype = fields.Str(required=False, validate=validate.OneOf(METADATA_TYPES))
+    meta = fields.Dict(missing=dict)
+
+
+def upgrade():  # noqa
+    column_schema = TableColumnSchema()
+    row_schema = TableRowSchema()
+    op.add_column(
+        'csv_file',
+        sa.Column('column_json', sqlalchemy_utils.types.json.JSONType(), nullable=True)
+    )
+    op.add_column(
+        'csv_file',
+        sa.Column('row_json', sqlalchemy_utils.types.json.JSONType(), nullable=True)
+    )
 
     # move row/column information into csv_file table here:
+    for csv_file in CSVFile.query:
+        try:
+            column_json = column_schema.dump(csv_file.columns, many=True)
+            row_json = row_schema.dump(csv_file.rows, many=True)
 
[…]
