### 2020-07-06 f3aee50
Fix linting errors
