+++ e2e-test/jest.config.js
@@ -5,13 +5,8 @@ module.exports = {
   setupFilesAfterEnv: [
     // Normally, expect-puppeteer is specified by the jest-puppeteer preset
     'expect-puppeteer',
-    './jest-puppeteer-vuetify',
+    'jest-puppeteer-vuetify',
     './jest.setup.js',
-  ],
-
-  transformIgnorePatterns: [
-    // Apply Babel transformation to jest-puppeteer-vuetify
-    'node_modules/(?!jest-puppeteer-vuetify/)',
   ],
 
   // extend test timeout to 1 hour when debugging
