+++ e2e-test/src/tests/load.test.js
@@ -1,28 +1,17 @@
-// import { CLIENT_URL } from '../util';
-
 import { vBtn } from 'jest-puppeteer-vuetify';
-import { toClickXPath } from 'jest-puppeteer-vuetify/src/jest-xpaths';
+import { CLIENT_URL } from '../util';
 
-  describe('home page ', () => {
-    it('load home page', async () => {
-      await Promise.all([
-        //
-        //  page.goto(CLIENT_URL),
-        page.goto('http://localhost:8080'),
-        page.waitForNavigation({ waitUntil: 'networkidle0' }),
-      ]);
-    });
 
-    // propably need to use the jest-puppeteer-vuetify helper
-    it('click button', async () => {
-      // await expect(page).toClick('button', { text: 'Your Datasets' });
-      // toClickXPath
-      await expect(page).toClickXPath(vBtn('Your Datasets'));
-    });
+describe('home page', () => {
+  // eslint-disable-next-line jest/expect-expect
+  it('load home page', async () => {
+    await Promise.all([
+      page.goto(CLIENT_URL),
+      page.waitForNavigation({ waitUntil: 'networkidle0' }),
+    ]);
   });
 
-
-  // await Promise.all([
-  //   expect(page).toClickXPath(vBtn('Your Datasets')),
-  //   page.waitForNavigation({ waitUntil: 'networkidle0' }),
-  // ]);
+  it('click button', async () => {
+    await expect(page).toClickXPath(vBtn('Your Datasets'));
+  });
+});
