### 2020-07-06 e06594f
Minor fixes
