+++ e2e-test/src/util.js
@@ -1,1 +1,2 @@
+/* eslint-disable import/prefer-default-export */
 export const { CLIENT_URL } = process.env;
