+++ e2e-test/README.md
@@ -14,10 +14,10 @@ Be aware of this when setting up CI or Docker images though.
 
 ## Running Tests
 You will need a running instance of the app, both the `web` and `flask` components.
-Assuming the web app is running at `http://localhost:8080`:
+Assuming the web app is running at `http://localhost:8085`:
 ```bash
 # within "test"
-CLIENT_URL=http://localhost:8080 yarn run test
+CLIENT_URL=http://localhost:8085 yarn run test
 ```
 
 ## Debugging Tests
@@ -26,7 +26,7 @@ For browser based tests, it is very helpful to be able to see the state of the b
 Use this command to run the browser in headful mode and extend the Jest test timeout to 1 hour:
 ```bash
 # within "test"
-CLIENT_URL=http://localhost:8080 yarn run test-debug
+CLIENT_URL=http://localhost:8085 yarn run test-debug
 ```
 You can also include `await jestPuppeteer.debug();` at any point in the test to create a manual breakpoint.
 
@@ -54,63 +54,3 @@ For example:
 await expect(page).toFill('input#username', 'DandiDan');
 await expect(page).toClick('input#submit`);
 ```
-
-### jest-xpaths
-Puppeteer and `expect-puppeteer` prefer using CSS selectors whenever possible.
-They do have some support for XPath locators, but they cannot be used for everything that CSS selectors can be used for.
-Sadly, many elements generated by Vuetify cannot practically be identified with CSS selectors, so sometimes XPaths are required.
-To help remedy this, the `jest-xpaths` module registers several new methods on the Jest `expect` object.
-For instance:
-```javascript
-await expect(page).toFillXPath('//input[@id="username"]', 'DandiDan');
-await expect(page).toClickXPath('//input[@id="submit"]');
-```
-This module is still a work in progress.
-If you think it should be capable of something that it doesn't do yet, feel free to augment it.
-
-### vuetify-xpaths
-Despite XPaths sometimes being the only option for locating Vuetify elements, they can still be painful to write and maintain.
-To help with this, we have the `vuetify-xpaths` helper which generates XPaths (element locators) for common Vuetify elements.
-For example, to fill a text field labelled `Username` and click a button labelled `Submit`, you can do this:
-```javascript
-await expect(page).toFillXPath(vTextField('Username'), 'DandiDan');
-await expect(page).toClickXPath(vBtn('Submit'));
-```
-#### Usage
-The canonical argument to any `vElement` is a destructured object.
-The arguments are unique to each element, but generally will have a `content` argument and a `cssClass` argument.
-If the argument is not an object, it is assumed to be `content`.
-* If `content` is falsy or absent, like `vFoo()` or `vFoo({content: null})`, the XPath will match any `v-foo`.
-* If `content` is a string, like `vFoo('Hello')` or `vFoo({content: 'World'})`, the XPath will match any `v-foo` which contains the string.
-* If `content` is a valid XPath, like `vFoo(vBar())`, the XPath will match any `v-foo` which contains a `v-bar`.
-Note that `vBar` is not necessarily an immediate child of `vFoo`; `vFoo(vBar())` will also match `<v-foo><div><v-bar /></div></v-foo>`.
-* If `content` is an array, every element in the array is treated as a separate `content`. `vFoo(['Hello', vBar()])` will match any `v-foo` which contains both a `v-bar` element and the text `Hello`.
-
-The same assumptions are generally made for different arguments.
-Here are some examples selectors and DOM elements they will locate:
-```
-vFoo({ contents: vBar('Hello World!'), cssClass: ['world', 'hello'] })
-
-<v-foo class='hello world'>
-  <v-bar>
-    Hello World!
-  </v-bar>
-</v-foo>
-```
-
-The second argument will generally be an object whose options will vary based on the element.
[…]
