### 2020-07-10 71c9fad
Add an endpoint for ROC analysis
