### 2020-07-14 e75d0d1
show mouse cursor on line, rename analysis Renames "ROC Curve" to "Group Prediction"
