### 2020-07-14 530e90e
Return area under curve for roc
