### 2020-07-15 76de144
clean up getting roc data from endpoint Use existing mechanisms to get the data instead of directly using axios
