### 2020-07-17 43db3d3
add additional analyses for ROC curve (see below) - Update roc endpoint with new script to allow for multiple metabolites - Implement factor analysis R script as an endpoint - Adjustments to web client to account for the new endpoints - Use raw html instead of pug for RocCurveTile component
