+++ devops/viime/R/analyses.R
@@ -156,31 +156,35 @@ clustered_heatmap <- function(measurements) {
 #' roc_analysis
 #'
 #' @export
-roc_analysis <- function(measurements, groups, group_name, column_name, method) {
+roc_analysis <- function(measurements, groups, group1_name, group2_name, column_names, method) {
   library(pROC)
   library(randomForest)
+  df <- read.csv(measurements, row.names=1, check.names=TRUE)
+  groups <- read.csv(groups, row.names=1, check.names=TRUE)
+  groups <- as.factor(groups[, 1])
 
-  df = read.csv(measurements, row.names=1, check.names=FALSE)
-  groups = read.csv(groups, row.names=1, check.names=FALSE)
-  groups = as.factor(groups[, 1])
-  group_mask = ifelse(groups == group_name, 1, 0)
+  group_mask <- numeric() # intialize empty numeric vector
+  for (i in groups) {
+    if (i == group1_name) {
+        group_mask <- c(group_mask, 0)
+    } else if (i == group2_name) {
+        group_mask <- c(group_mask, 1)
+    }
+  }
 
-  #############
-  #Get predicted values
-
-  column = df[[ column_name ]];
+  column_names <- read.csv(column_names, row.names=1, check.names=TRUE)
+  c <- unlist(column_names) # convert dataframe to a vector
+  c <- make.names(c) # convert column names to valid R names
+  columns <- df[c] # only select the columns/metabolites we want
+  new <- cbind(group_mask, columns)
 
   if (method == "logistic_regression") {
     #-# Logistic Regression
-    glm.fit = glm(group_mask ~ column, family=binomial)
-
-    # Save predicted values from logistic regression
+    glm.fit <- glm(group_mask ~ . , family=binomial, data=new)
     pred <- glm.fit$fitted.values
   } else if (method == "random_forest") {
     #-# Random forest
-    rf.model <- randomForest::randomForest(factor(group_mask)~column)
-
-    # Save predicted values from random forest
+    rf.model <- randomForest::randomForest(factor(group_mask)~ ., data=new)
     pred <- rf.model$votes[,1]
   } else {
     stop("Invalid method")
@@ -195,4 +199,59 @@ roc_analysis <- function(measurements, groups, group_name, column_name, method) 
       thresholds=roc.a$thresholds,
       auc=rep(roc.a$auc, length(roc.a$thresholds))
   )
+}
+
+#' factor_analysis
+#'
+#' @export
+factor_analysis <- function(measurements, threshold) {
+  library(psych) 
+
+  m.df <- read.csv(measurements, row.names=1, check.names=FALSE)
+
+  ##########################################
+  # Code for factor analysis
+
+
+  #-#-#- Defining number of factors we need -#-#-#
+  #- (Eigenvalues higher than 1) -#
+
+  # Principal Component Analysis to get eigenvalues
+  pca_a <- prcomp(m.df, center=T, scale=T)
+
+  #getting the eigenvalues
+  MS.eig <- (pca_a$sdev)^2 
+
+  #Eigenvalues higher than 1 to see how many factors we need
+  ncomp <- sum(MS.eig >= 1)
+
+  ##############################
+  #-#-#- Factor Analysis -#-#-#
+
+  # Factor analysis
+  #Using ncomp, number of components with eigenvalue higher than 1
+  fitpsy <- psych::fa(m.df, nfactors=ncomp, rotate="varimax") 
+
+  #Save some results from factor analysis
+  eigen.values <- fitpsy$e.values #eigenvalues
+  loadings <- fitpsy$Structure[,1:ncomp] #Loadings
[…]
