### 2020-07-17 b3bf427
loading threshold slider added, update roc test - Add a slider to the UI to change threshold used in factor analysis - Update roc analysis test to account for changes - Remove some CSS classes that are no longer used
