### 2020-07-20 c5a8e97
fix bug with groups in roc analysis Fixed swapping group 1 and 2 resulting in a different ROC curve
