### 2020-07-21 #2410 @joannawolthuis
Hi @lpantano ! I'm taking a look - but here's a perhaps noob question: this is my first review, should I be writing a more extensive document (and if so, where should I submit it?) using the checklist as a backbone/guide? Or is this a very modern type of review where it is just the checklist? :)
