### 2020-07-22 #2410 @lpantano
Hi @joannawolthuis, you don't need to write an extensive review. If you want to comment on something, you can do it. But the important it all the items in the checklist are checked. If not, you can open an issue in the tool repository to address the issue, so they can fix it and you can mark all the items in the checklist.
