### 2020-07-22 e6b719b
add some code comments and rename roc component - Renamed roc tile component to be consistent w/ existing vis components - Add some more comments documenting code in GroupPredictionTile
