### 2020-07-24 bf80b61
move R dependencies to base docker image
