+++ devops/viime/DESCRIPTION
@@ -4,7 +4,7 @@ Version: 0.0.0.9000
 Authors@R: person("First", "Last", email = "first.last@example.com", role = c("aut", "cre"))
 Description: The package contains functions that are exposed via RPC for the Metabolomics web application.
 Depends: R (>= 3.5.2)
-Imports: car, emmeans, pROC, randomForest
+Imports: car, emmeans, randomForest
 License: Apache 2.0
 Encoding: UTF-8
 LazyData: true
