### 2020-07-27 f54069d
Add ROC analysis/visualization (#583)
* Add an endpoint for ROC analysis

* Return area under curve for roc

* initial ROC curve

* add selector for method

* add dropdown for group

* fix bug with metabolites with spaces

* use 1 - specificity for x-axis

* show mouse cursor on line, rename analysis
Renames "ROC Curve" to "Group Prediction"

* clean up getting roc data from endpoint
Use existing mechanisms to get the data instead of directly using axios

* use auc value returned by the roc endpoint

* import individual d3 functions

* option to limit metabolites by global selection

* add add
[…]
