+++ devops/viime/R/analyses.R
@@ -3,8 +3,8 @@
 #'
 #' @export
 wilcoxon_test_z_scores <- function(measurements, groups, log_transformed=FALSE) {
-  Metab = read.csv(measurements, row.names=1)
-  groups = read.csv(groups, row.names=1)
+  Metab = read.csv(measurements, row.names=1, check.names=FALSE)
+  groups = read.csv(groups, row.names=1, check.names=FALSE)
 
   # take the first column
   Group = as.factor(groups[, 1])
@@ -64,8 +64,8 @@ wilcoxon_test_z_scores <- function(measurements, groups, log_transformed=FALSE) 
 #'
 #' @export
 anova_tukey_adjustment <- function(measurements, groups, log_transformed=FALSE) {
-  Metab = read.csv(measurements, row.names=1)
-  groups = read.csv(groups, row.names=1)
+  Metab = read.csv(measurements, row.names=1, check.names=FALSE)
+  groups = read.csv(groups, row.names=1, check.names=FALSE)
 
   # take the first column
   Group = as.factor(groups[, 1])
@@ -151,4 +151,116 @@ clustered_heatmap <- function(measurements) {
   close(f)
 
   list(data=OUT_CSV, row=unclass(ddr), col=unclass(ddc))
+}
+
+#' roc_analysis
+#'
+#' @export
+roc_analysis <- function(measurements, groups, group1_name, group2_name, column_names, method) {
+  library(pROC)
+  library(randomForest)
+  df <- read.csv(measurements, row.names=1, check.names=TRUE)
+  groups <- read.csv(groups, row.names=1, check.names=TRUE)
+  groups <- as.factor(groups[, 1])
+
+  group_mask <- numeric() # intialize empty numeric vector
+  for (i in groups) {
+    if (i == group1_name) {
+        group_mask <- c(group_mask, 0)
+    } else if (i == group2_name) {
+        group_mask <- c(group_mask, 1)
+    }
+  }
+  
+  # make sure row order is preserved
+  if (group_mask[1] == 0) {
+    # remove all rows not in group1 or group2
+    df <- rbind(df[groups == group1_name,], df[groups == group2_name,])
+  } else {
+    df <- rbind(df[groups == group2_name,], df[groups == group1_name,])
+  }
+
+
+  column_names <- read.csv(column_names, row.names=1, check.names=TRUE)
+  c <- unlist(column_names) # convert dataframe to a vector
+  c <- make.names(c) # convert column names to valid R names
+  columns <- df[c] # only select the columns/metabolites we want
+  new <- cbind(group_mask, columns)
+
+  if (method == "logistic_regression") {
+    #-# Logistic Regression
+    glm.fit <- glm(group_mask ~ . , family=binomial, data=new)
+    pred <- glm.fit$fitted.values
+  } else if (method == "random_forest") {
+    #-# Random forest
+    rf.model <- randomForest::randomForest(factor(group_mask)~ ., data=new)
+    pred <- rf.model$votes[,1]
+  } else {
+    stop("Invalid method")
+  }
+
+  #############
+  #ROC analysis
+  roc.a <- pROC::roc(group_mask, pred, plot=FALSE)
+  out = data.frame(
+      sensitivities=roc.a$sensitivities,
+      specificities=roc.a$specificities,
+      thresholds=roc.a$thresholds,
+      auc=rep(roc.a$auc, length(roc.a$thresholds))
+  )
+}
+
+#' factor_analysis
+#'
+#' @export
+factor_analysis <- function(measurements, threshold) {
+  library(psych) 
+
+  m.df <- read.csv(measurements, row.names=1, check.names=FALSE)
+
[…]
