### 2020-07-27 d575ec0
catch errors from failed factor analysis in R code
