### 2020-07-27 7ca2b92
check for empty column array in roc analysis
