### 2020-07-27 b456bfd
Catch R errors in ROC analysis (#594)
* catch errors from failed factor analysis in R code

* check for empty column array in roc analysis
