### 2020-07-28 bbfdecb
Adding delete end-to-end tests
Also rearrange how end-to-end test files are stored
