+++ e2e-test/src/tests/delete.test.js
@@ -3,8 +3,9 @@ import {
   vCard,
   vIcon,
   vListTile,
+  vList,
 } from 'jest-puppeteer-vuetify';
-import { CLIENT_URL, deleteAllDatasets, allCleared } from '../util';
+import { CLIENT_URL, deleteAllDatasets } from '../util';
 
 
 describe('delete dataset', () => {
@@ -23,12 +24,13 @@ describe('delete dataset', () => {
     await page.waitForXPath(vListTile({ title: 'deleteTest.csv ' }));
     await page.waitForXPath(vListTile({ title: 'deleteTest.csv ', content: 'Dataset ready for analysis' }));
 
+    await new Promise((resolve) => setTimeout(resolve, 1000));
     await expect(page).toClickXPath(vBtn(vIcon('mdi-close')));
     await page.waitForXPath(vCard({ headline: 'Really delete 1 dataset?' }));
     await expect(page).toClickXPath(vBtn('Delete'));
 
-    const previouslyUploadedFile = await expect(page).toContainXPath(vListTile({ title: 'deleteTest.csv ' }));
-    await expect(previouslyUploadedFile).toBeUndefined();
+    await new Promise((resolve) => setTimeout(resolve, 1000));
+    await expect(page).not.toContainXPath(vListTile({ title: 'deleteTest.csv ' }));
   });
 
   it('clear all', async () => {
@@ -55,11 +57,13 @@ describe('delete dataset', () => {
     await page.waitForXPath(vListTile({ title: 'deleteTest3.csv ' }));
     await page.waitForXPath(vListTile({ title: 'deleteTest3.csv ', content: 'Dataset processed with 1 validation failures' }));
 
+    await new Promise((resolve) => setTimeout(resolve, 1000));
     await expect(page).toClickXPath(vBtn('clear all'));
-
     await page.waitForXPath(vCard({ headline: 'Really delete 3 datasets?' }));
     await expect(page).toClickXPath(vBtn('Delete'));
 
-    await allCleared();
+    await new Promise((resolve) => setTimeout(resolve, 1000));
+    const uploadedItems = await page.$x(vList(vListTile({ cssClass: 'theme--light' })));
+    await expect(uploadedItems).toHaveLength(0);
   });
 });
