+++ e2e-test/src/util.js
@@ -1,18 +1,13 @@
-import { vBtn, vList } from 'jest-puppeteer-vuetify';
+import { vBtn, vList, vListTile } from 'jest-puppeteer-vuetify';
 
 export const { CLIENT_URL } = process.env;
 
 export async function deleteAllDatasets() {
-  // firstUploadedItem returns a uploaded file node list
-  const firstUploadedItem = await page.$x(vList({ content: 'upload-list theme--light' }));
-  if (firstUploadedItem.length !== 0) {
+  // $x returns a list of all elements matching an XPath
+  const uploadedItems = await page.$x(vList(vListTile({ cssClass: 'theme--light' })));
+  if (uploadedItems.length > 0) {
     await expect(page).toClickXPath(vBtn('clear all'));
     await page.waitForXPath(vBtn('Delete'));
     await expect(page).toClickXPath(vBtn('Delete'));
   }
-}
-
-export async function allCleared() {
-  const firstUploadedItem = await page.$x(vList({ content: 'upload-list theme--light' }));
-  await expect(firstUploadedItem).toHaveLength(0);
 }
