### 2020-07-29 2aebd89
Added neccessary comments
