### 2020-07-30 686440f
3 clean up table realted tests
