### 2020-07-30 e005d9a
Add usePlotData and convert ScreePlotTile to TS
