### 2020-08-02 #2410 @carleshf
Hi!

Sorry for the delay, I'm happy to review the tool too if I'm not too late.

*Carles Hernández*

Missatge de Roni Choudhury <notifications@github.com> del dia dl., 27 de
jul. 2020 a les 23:59:

> Great, thanks for confirming!
>
