### 2020-08-03 0f435c3
add error handling to ANOVA test Terminates the ANOVA test gracefully instead of crashing when it fails
