+++ viime/analyses.py
@@ -39,10 +39,15 @@ def anova_test(measurements: pd.DataFrame, groups: pd.Series,
 
     data = data.rename(columns={'(Intercept)': 'Intercept'})
 
+    formatted_data = clean(data).to_dict(orient='records')
+
+    if formatted_data[0].get('error') is not None:
+        return { 'error': formatted_data[0].get('error') }
+
     return {
         'groups': list(set(groups)),
         'pairs': [v for v in list(data)[4:] if not v.endswith('FoldChange')],
-        'data': clean(data).to_dict(orient='records')
+        'data': formatted_data
     }
 
 
