### 2020-08-04 #2410 @joannawolthuis
Hi @lpantano! Same as @jkanche, have opened some issues. One question - for the 'installation' checkmark, does that still count if this aimed to be a webserver? If so I need to ask them to clarify their installation procedure..
