### 2020-08-04 d488b80
Review fixes
- Fix typo in `usePlotData.ts`
- Properly destructure args in `usePlotData.ts` `watchEffect`
- Add comment for temporary Vue 2 hack in `ScreePlotTile.vue`
