### 2020-08-04 9f9d862
code clean up, rename and make upload a util function
