### 2020-08-05 #2410 @jkanche
I agree with @joannawolthuis, I would also like to see more detailed installation instructions. I am confused on how one can setup an instance Viime for their datasets.

May be provide a docker container or instructions on how one can set everything up will be good.
