### 2020-08-05 #2410 @lpantano
Thank you for the updates @joannawolthuis, @jkanche. Yes, the installation needs to be addressed as well for servers. It is a very good point.
