### 2020-08-05 70c1ae3
3 test cases for transform table and analyze table buttons/checkboxes
