### 2020-08-05 girder/viime#611 @waxlamp
Clarify setup instructions
The JOSS reviewers (@joannawolthuis and @jkanche) have both requested improved setup instructions (see https://github.com/openjournals/joss-reviews/issues/2410#issuecomment-668530730 and https://github.com/openjournals/joss-reviews/issues/2410#issuecomment-668917059).

@joannawolthuis, @jkanche: would you mind providing some detail of where things went wrong for you in following the setup procedure in the [README](https://github.com/girder/viime#viime)?
