### 2020-08-06 f504436
Add error handling to ANOVA (#605)
* Catch ANOVA test errors on the server and return an error message in the response instead of crashing opencpu server

* Display the error message on the client instead of just a blank screen
