### 2020-08-06 23066c1
Set up docker-compose for server/db (#579)
