### 2020-08-07 e70a445
add r dependency to base dockerfile
