### 2020-08-07 caf686a
Convert PcaPage to use Typescript
Also convert the three components on the PcaPage, ScorePlot, ScreePlot,
and LoadingsPlot.

Using the composition API on PcaPage created a circular dependency:
```
PcaPage.js -> usePlotData.ts -> store/index.js -> analyses.js ->
PcaPage.js
```

The problem was that `analyses.js` provides a list of all available
analyses and all data required to describe them, including the
components that should be used to render them. `store/index.js` imported
`analyses.js` to dynamically determine what analyses needed plot data
stores, which had the side affect of importing a
[…]
