+++ web/src/components/vis/analyses.js
@@ -7,8 +7,6 @@ import CorrelationTile from './CorrelationTile.vue';
 import PcaPage from './PcaPage/PcaPage.vue';
 import BoxPlotLargeTile from './BoxPlotLargeTile.vue';
 import GroupPredictionTile from './GroupPredictionTile.vue';
-import { plot_types } from '../../utils/constants';
-import { correlation_methods } from './constants';
 import vuetify from '../../utils/vuetifyConfig';
 
 export default [
@@ -18,8 +16,6 @@ export default [
     shortName: 'PCA',
     description: 'TODO',
     component: PcaPage,
-    args: {},
-    type: plot_types.ANALYSIS,
     icon: vuetify.icons.pca,
   },
   {
@@ -33,11 +29,11 @@ export default [
           using a series of box plots.
         </p>
         <p>Each metabolite appears along the y-axis, with a horizontal box plot showing
-          the four quartile values, emphasizing
-          the interquartile range (IQR) with solid bars. Individual outliers appear as well:
-          normal ones, falling at
-          least 1.5 IQRs away from the interquartile range, as dots; and "far-out" outliers,
-          falling at least 3 IQRs away from the interquartile range, as larger dots.
+        the four quartile values, emphasizing
+        the interquartile range (IQR) with solid bars. Individual outliers appear as well:
+        normal ones, falling at
+        least 1.5 IQRs away from the interquartile range, as dots; and "far-out" outliers,
+        falling at least 3 IQRs away from the interquartile range, as larger dots.
         </p>
         <p>
           Hovering the mouse pointer over various parts of the plot will show detailed
@@ -47,8 +43,6 @@ export default [
       </div>);
     },
     component: BoxPlotLargeTile,
-    args: {},
-    type: plot_types.ANALYSIS,
     icon: vuetify.icons.boxplot,
     iconStyle: {
       transform: 'rotate(90deg)scale(-1,1)',
@@ -80,8 +74,6 @@ export default [
       </div>);
     },
     component: WilcoxonPlotTile,
-    args: {},
-    type: plot_types.ANALYSIS,
     icon: vuetify.icons.metaboliteTable,
   },
   {
@@ -94,8 +86,6 @@ export default [
       </div>);
     },
     component: WilcoxonVolcanoPlotTile,
-    args: {},
-    type: plot_types.ANALYSIS,
     icon: vuetify.icons.pca,
   },
   {
@@ -104,8 +94,6 @@ export default [
     shortName: 'ANOVA',
     description: 'Test to compare 3 or more groups assuming normal distribution, the group pairwise comparisons are adjusted with Tukey HSD',
     component: AnovaTableTile,
-    args: {},
-    type: plot_types.ANALYSIS,
     icon: vuetify.icons.metaboliteTable,
   },
   {
@@ -114,8 +102,6 @@ export default [
     shortName: 'ANOVA Volcano Plot',
     description: 'Test to compare 3 or more groups assuming normal distribution, the group pairwise comparisons are adjusted with Tukey HSD',
     component: AnovaVolcanoPlotTile,
-    args: {},
-    type: plot_types.ANALYSIS,
     icon: vuetify.icons.pca,
   },
   {
@@ -124,13 +110,6 @@ export default [
     shortName: 'Heatmap',
     description: 'Is a graphical representation of the concentration differences between variables and samples',
     component: HeatmapTile,
-    args: {
-      column: null,
-      column_filter: null,
-      row: null,
-      row_filter: null,
-    },
-    type: plot_types.ANALYSIS,
     icon: vuetify.icons.heatmap,
   },
   {
@@ -141,11 +120,6 @@ export default [
[…]
