### 2020-08-10 girder/viime#611 @waxlamp
Hi @joannawolthuis and @jkanche! This [new section of the README](https://github.com/girder/viime#getting-started-docker-compose), explains how to spin up the server backend locally using Docker. Please check it out and let us know if it works for you.
