### 2020-08-11 a0fa97f
change endpoint to work with `plotData` mixin
