### 2020-08-11 de07c7f
finish pls-da loadings plot
