+++ viime/analyses.py
@@ -118,7 +118,7 @@ def factor_analysis(measurements: pd.DataFrame, threshold=0.4) -> Dict[str, List
     return clean(data).to_dict(orient='list')
 
 
-def plsda(measurements: pd.DataFrame, groups: pd.DataFrame, num_of_components=5, mode="scores"):
+def plsda(measurements: pd.DataFrame, groups: pd.DataFrame, num_of_components: int, mode='scores'):
     files = {
         'measurements': measurements.to_csv().encode(),
         'groups': groups.to_csv(header=True).encode()
