### 2020-08-11 e0107cc
get plsda scores plot to display correctly
