### 2020-08-11 21f9ab5
restructure the test cases and files
