### 2020-08-12 #2410 @joannawolthuis
Just checking in, thanks for your patience! I have opened some more issues. The installation works well now, just have some issues on startup in terms of example datasets etc. (opened an issue on their github)
Just another question - they are currently dealing with some bugs, as one always has with a tool, new updates will always add new bugs - should i only check off the mark for expected/promised functionality once they have ironed those out? (some plots are just not showing at all, etc).
