### 2020-08-12 #2410 @joannawolthuis
Oh, one more - is there a supplemental info or something for this paper? Otherwise they have no comparison at all with other tools/state of the art, should I address that on their Github page or through another channel?
