### 2020-08-12 girder/viime#624 @joannawolthuis
Comparison to other tools
Hi there! In the context of the JOSS review, a comparison with other tools is required. I couldn't find one in the manuscript, so if there isn't a supplemental that I missed, I'd like to request such a comparison (perhaps with one of those cliche comparison tables :) )
