### 2020-08-12 c293af4
fix plsda bug default number of components for plsda was being computed incorrectly
