### 2020-08-13 969334c
handle num_of_components being too large
