### 2020-08-13 6f27c15
add test for plsda analysis
