### 2020-08-13 e6225c7
fix plsda test and add sample datasets for it
