+++ tests/test_analyses.py
@@ -77,80 +77,76 @@ def test_factor_analysis(client, test_dataset):
         assert 'variances' not in resp.json
 
 def test_plsda(client):
-    with (Path(__file__).parent / 'roc.csv').open('rb') as f:
-        csv_data = f.read()
-    data = {'file': (BytesIO(csv_data), 'roc.csv', 'text/csv')}
-    resp = client.post(
-        url_for('csv.upload_csv_file'), data=data, content_type='multipart/form-data')
-    assert resp.status_code == 201
-    csv_id = resp.json['id']
+    for i in range(2):  # run test on two datasets
+        with (Path(__file__).parent / f'plsda/plsda{i+1}.csv').open('rb') as f:
+            csv_data = f.read()
+        data = {'file': (BytesIO(csv_data), f'plsda/plsda{i+1}.csv', 'text/csv')}
+        resp = client.post(
+            url_for('csv.upload_csv_file'), data=data, content_type='multipart/form-data')
+        assert resp.status_code == 201
+        csv_id = resp.json['id']
 
-    data = {'changes': [
-        {'context': 'column', 'index': 71, 'label': 'group'},
-        {'context': 'column', 'index': 1, 'label': 'measurement'}
-    ]}
-    resp = client.put(url_for('csv.batch_modify_label', csv_id=csv_id), json=data)
-    assert resp.status_code == 200
+        resp = client.post(url_for('csv.save_validated_csv_file', csv_id=csv_id))
+        assert resp.status_code == 201
 
-    resp = client.post(url_for('csv.save_validated_csv_file', csv_id=csv_id))
-    assert resp.status_code == 201
+        data = {}  # no data
+        resp = client.get(url_for('csv.get_plsda', csv_id=str(csv_id)), json=data)
+        assert resp.status_code == 200
+        keys = {'scores', 'loadings'}
+        assert keys == set(resp.json.keys())
 
-    data = {}  # no data
-    resp = client.get(url_for('csv.get_plsda', csv_id=str(csv_id)), json=data)
-    assert resp.status_code == 200
-    keys = {'scores', 'loadings'}
-    assert keys == set(resp.json.keys())
+        columns = client.get(url_for('csv.list_columns', csv_id=str(csv_id))).json
+        rows = client.get(url_for('csv.list_rows', csv_id=str(csv_id))).json
+        columns = [
+            column['column_header'] for column in columns if column['column_type'] == 'measurement'
+        ]
+        rows = [row for row in rows if row['row_type'] == 'sample']
 
-    columns = client.get(url_for('csv.list_columns', csv_id=str(csv_id))).json
-    rows = client.get(url_for('csv.list_rows', csv_id=str(csv_id))).json
-    columns = [
-        column['column_header'] for column in columns if column['column_type'] == 'measurement'
-    ]
-    rows = [row for row in rows if row['row_type'] == 'sample']
+        # number of components should be equal to min(columns, rows) by default
+        num_of_components = min(len(columns), len(rows))
 
-    # number of components should be equal to min(columns, rows) by default
-    num_of_components = min(len(columns), len(rows))
+        loadings = resp.json.get('loadings')
+        for loading in loadings:
+            assert loading.get('col') in columns
+            assert loading.get('loadings') is not None
 
-    loadings = resp.json.get('loadings')
-    for loading in loadings:
-        assert loading.get('col') in columns
-        assert loading.get('loadings') is not None
+            assert len(loading.get('loadings')) == num_of_components
 
-        assert len(loading.get('loadings')) == num_of_components
+        scores = resp.json.get('scores')
 
-    scores = resp.json.get('scores')
+        assert 'sdev' in scores.keys()
[…]
