### 2020-08-13 0b164a4
add some code comments to explain test
