### 2020-08-13 5293758
add test for factor analysis
