+++ tests/test_analyses.py
@@ -5,25 +5,32 @@ from pathlib import Path
 from flask import url_for
 
 
-def test_roc(client):
-    with (Path(__file__).parent / 'roc.csv').open('rb') as f:
+def upload_test_dataset(client, file, changes=None):
+    with (Path(__file__).parent / file).open('rb') as f:
         csv_data = f.read()
-    data = {'file': (BytesIO(csv_data), 'roc.csv', 'text/csv')}
+    data = {'file': (BytesIO(csv_data), file, 'text/csv')}
     resp = client.post(
         url_for('csv.upload_csv_file'), data=data, content_type='multipart/form-data')
     assert resp.status_code == 201
     csv_id = resp.json['id']
 
-    data = {'changes': [
-        {'context': 'column', 'index': 71, 'label': 'group'},
-        {'context': 'column', 'index': 1, 'label': 'measurement'}
-    ]}
-    resp = client.put(url_for('csv.batch_modify_label', csv_id=csv_id), json=data)
-    assert resp.status_code == 200
+    if changes is not None:
+        data = {'changes': changes}
+        resp = client.put(url_for('csv.batch_modify_label', csv_id=csv_id), json=data)
+        assert resp.status_code == 200
 
     resp = client.post(url_for('csv.save_validated_csv_file', csv_id=csv_id))
     assert resp.status_code == 201
 
+    return csv_id
+
+
+def test_roc(client):
+    changes = [
+        {'context': 'column', 'index': 71, 'label': 'group'},
+        {'context': 'column', 'index': 1, 'label': 'measurement'}
+    ]
+    csv_id = upload_test_dataset(client, 'roc.csv', changes)
     data = {
         'group1': 'C',
         'group2': 'H',
@@ -37,3 +44,33 @@ def test_roc(client):
     for key in keys:
         for value in resp.json[key]:
             assert value in {'Inf', '-Inf'} or 0 <= value <= 1
+
+
+def test_factor_analysis(client):
+    changes = [
+        {'context': 'column', 'index': 71, 'label': 'group'},
+        {'context': 'column', 'index': 1, 'label': 'measurement'}
+    ]
+
+    csv_id = upload_test_dataset(client, 'roc.csv', changes)
+
+    resp = client.get(url_for('csv.get_factors', csv_id=str(csv_id)))
+    assert resp.status_code == 200
+
+    eigenvalues = resp.json.get('eigenvalues')
+    factors = resp.json.get('factor')
+    metabolites = resp.json.get('metabolites')
+    variances = resp.json.get('variances')
+
+    assert len(eigenvalues) == len(factors) == len(metabolites) == len(variances)
+
+    for factor in factors:
+        assert isinstance(factor, int)
+
+    # test factor analysis that fails
+    resp = client.get(url_for('csv.get_factors', csv_id=str(csv_id), threshold=0.6))
+    assert resp.status_code == 200
+    assert 'eigenvalues' not in resp.json
+    assert 'factor' not in resp.json
+    assert 'metabolites' not in resp.json
+    assert 'variances' not in resp.json
