### 2020-08-13 181813b
Remove mention of ldc from README
