### 2020-08-13 63b8826
Remove lingering reference to LDC
