### 2020-08-14 e2c2315
implement Lilian's Wilcoxon test fix
