### 2020-08-17 0365c2d
catch wilcox error in R code
