### 2020-08-18 5e14760
use pytest fixture for test function
