### 2020-08-18 625139a
move fixture to `test_analyses.py`
