### 2020-08-18 e31b62b
Add test for factor analysis (#629)
