### 2020-08-18 38cb6bc
update plsda test to use new fixture
