### 2020-08-18 ed4451d
use append instead of preallocating lists
