### 2020-08-18 f7b2c67
use `maybeData`, send labels/rows back in plsda endpoint
