### 2020-08-18 60c37a3
fix plsda test to account for changes
