+++ viime/views.py
@@ -774,7 +774,11 @@ def get_plsda(validated_table: ValidatedMetaboliteTable, num_of_components: Opti
     labels = clean(pandas.concat([groups, labels], axis=1)).to_dict('list')
     rows = measurements.index.tolist()
 
-    return jsonify({'scores': formatted_scores, 'loadings': formatted_loadings, 'labels': labels, 'rows': rows})
+    return jsonify({
+        'scores': formatted_scores,
+        'loadings': formatted_loadings,
+        'labels': labels,
+        'rows': rows})
 
 
 def _group_test(method: Callable, validated_table: ValidatedMetaboliteTable,
