### 2020-08-18 ffeb4ee
return vip scores in plsda endpoint
