### 2020-08-19 girder/viime#611 @jkanche
cool, thanks for updating the instructions. Is there a reason why the web client cannot be a docker container and has to be run separately ?
