### 2020-08-19 girder/viime#611 @waxlamp
> cool, thanks for updating the instructions. Is there a reason why the web client cannot be a docker container and has to be run separately ?

I don't think there's anything preventing it, but since the web client build is a standard NPM workflow, we didn't feel a need to do it.

If you run `yarn install` followed by `yarn serve` from the `web/` directory, things should go pretty smoothly. Are you running into any trouble with that process?
