### 2020-08-19 girder/viime#611 @jkanche
No I was able to run everything locally. I thought since everything is already setup using docker, It might make it easier to include the client as part of the docker rather than have to do an extra step. but this works as wel
