### 2020-08-19 girder/viime#611 @waxlamp
Ah, I see. That idea does make sense and I've opened #637 to track it.

I will go ahead and close this issue since things seem to be working better than before. If you run into new problems, feel free to re-open.
