### 2020-08-21 7bce4f6
Add a help panel to VisTileLarge
This works by cribbing the description of a given analysis for use
within the VisTileLarge context.
