### 2020-08-21 e88f05f
Add description for PLSDA
