### 2020-08-24 86f2e12
Use better JS style in defining analysisMap
Following recommendations from @subdavis via
https://github.com/girder/viime/pull/644#discussion_r475596763.
