### 2020-08-24 4204a74
Strip out columns with zero variance in factor analysis
