### 2020-08-24 0017cdf
Improve the help text for correlation network
