### 2020-08-24 3c49ea8
Add note to README about where to find the sample data files
