### 2020-08-24 17f8fb8
Use imperative reduction function
This is to forestall possible performance problems. See [this
comment](https://github.com/girder/viime/pull/644#discussion_r475755390)
for details.
