### 2020-08-24 cc58ea9
Use a relative link for sample data directory
