### 2020-08-25 6777e3e
Improve table structure inferred during ingest
