### 2020-08-27 e94f050
Simplify nginx config
