+++ e2e-test/src/tests/download.test.js
@@ -1,5 +1,5 @@
 import { vBtn, vListTile } from 'jest-puppeteer-vuetify';
-import { uploadDataset } from '../util';
+import { uploadDataset, analyzeTable } from '../util';
 
 
 describe('download dataset', () => {
@@ -50,11 +50,8 @@ describe('download dataset', () => {
     await uploadDataset('downloadTest');
     await expect(page).toClickXPath(vListTile({ title: 'downloadTest.csv ', content: 'Dataset ready for analysis.' }) + vBtn('View Data'));
 
-    // click analyze table
-    await expect(page).toClickXPath(vListTile({ title: 'Analyze Table' }));
-
-    // click ANOVA
-    await expect(page).toClickXPath(vListTile({ title: 'ANOVA' }));
+    // open ANOVA analysis
+    await analyzeTable('ANOVA');
 
     // select 7 rows: Alanine, Glutamine, Nicotinurate, Sarcosine, Aspartate, Glucose, Glutathione
     /* eslint prefer-template: "off" */
