### 2020-08-31 68156d5
Fix tests for different upload behavior
