### 2020-09-02 ad27c1e
Add citations
