+++ joss/paper.bib
@@ -57,3 +57,50 @@
   year={2013},
   pages={149–179},
 }
+
+@article{xia:2009,
+    author = {Xia, Jianguo and Psychogios, Nick and Young, Nelson and Wishart, David S.},
+    title = "{MetaboAnalyst: a web server for metabolomic data analysis and interpretation}",
+    journal = {Nucleic Acids Research},
+    volume = {37},
+    number = {suppl_2},
+    pages = {W652-W660},
+    year = {2009},
+    month = {05},
+    abstract = "{Metabolomics is a newly emerging field of ‘omics’ research that is concerned with characterizing large numbers of metabolites using NMR, chromatography and mass spectrometry. It is frequently used in biomarker identification and the metabolic profiling of cells, tissues or organisms.  The data processing challenges in metabolomics are quite unique and often require specialized (or expensive) data analysis software and a detailed knowledge of cheminformatics, bioinformatics and statistics. In an effort to simplify metabolomic data analysis while at the same time improving user accessibility, we have developed a freely accessible, easy-to-use web server for metabolomic data analysis called MetaboAnalyst. Fundamentally, MetaboAnalyst is a web-based metabolomic data processing tool not unlike many of today's web-based microarray analysis packages. It accepts a variety of input data (NMR peak lists, binned spectra, MS peak lists, compound/concentration data) in a wide variety of formats. It also offers a number of options for metabolomic data processing, data normalization, multivariate statistical analysis, graphing, metabolite identification and pathway mapping. In particular, MetaboAnalyst supports such techniques as: fold change analysis, t-tests, PCA, PLS-DA, hierarchical clustering and a number of more sophisticated statistical or machine learning methods. It also employs a large library of reference spectra to facilitate compound identification from most kinds of input spectra. MetaboAnalyst guides users through a step-by-step analysis pipeline using a variety of menus, information hyperlinks and check boxes.  Upon completion, the server generates a detailed report describing each method used, embedded with graphical and tabular outputs. MetaboAnalyst is capable of handling most kinds of metabolomic data and was designed to perform most of the common kinds of metabolomic data analyses.  MetaboAnalyst is accessible at http://www.metaboanalyst.ca}",
+    issn = {0305-1048},
+    doi = {10.1093/nar/gkp356},
+    url = {https://doi.org/10.1093/nar/gkp356},
+    eprint = {https://academic.oup.com/nar/article-pdf/37/suppl\_2/W652/3933058/gkp356.pdf},
+}
+
+@article{chong:2018,
+    author = {Chong, Jasmine and Soufan, Othman and Li, Carin and Caraus, Iurie and Li, Shuzhao and Bourque, Guillaume and Wishart, David S and Xia, Jianguo},
+    title = "{MetaboAnalyst 4.0: towards more transparent and integrative metabolomics analysis}",
+    journal = {Nucleic Acids Research},
+    volume = {46},
+    number = {W1},
+    pages = {W486-W494},
+    year = {2018},
+    month = {05},
[…]
