### 2020-09-02 girder/viime#624 @waxlamp
@joannawolthuis, we just added some commits to the paper text with some discussion of comparison to our closest "competitor". Please take a look and let me know what you think!
