### 2020-09-02 #2410 @waxlamp
@lpantano, @joannawolthuis, and @jkanche, thank you all for serving as reviewers and providing valuable feedback to us!

I've been trying to keep up with the various issues and requests, and I *believe* I've at least addressed everything that has been brought up so far. That is, we have fixed some bugs, made documentation updates to explain certain things better, and filed issues for other bugs. My goal is to get the software and paper up to review standards, since (unfortunately) we don't have the person-hours to fix everything, while filing issues for the rest (to be fixed once we have more funding/FTEs).

I am hoping to sync up with everyone by sending this message. When you get a chance, could you look things over and let me know what still needs to be done?

Thanks again for your time and effort in being our reviewers!
