### 2020-09-03 4dfd4a0
Add client image (#655)
* Add client image

* Simplify nginx config

Co-authored-by: Jonathan Beezley <jonathan.beezley@kitware.com>
Co-authored-by: Daniel Chiquito <daniel.chiquito@kitware.com>
