+++ devops/docker/start.sh
@@ -1,0 +1,2 @@
+set -e
+nginx -g 'daemon off;'
