### 2020-09-07 a47c28e
Turn sdev into a flat list by getting first element
