### 2020-09-08 girder/viime#624 @joannawolthuis
Hi @waxlamp, thanks for adding this section! I have one comment - please double check the facts on MetaboAnalyst, they do have more missing value imputation options than the 1/5 lowest value one, KNN means., replace with mean/median, PPCA etc (although no MCAR etc. estimation):
![image](https://user-images.githubusercontent.com/16805477/92458321-8165b880-f1c5-11ea-8a68-d70dde31dc34.png)

Secondly, maybe you could consider adding at least another 2 tools or so that are similar, I know there are other tools that compare themselves to MetaboAnalyst as well (WebSpecmine is one example that is R-based, and there are more, MzMine), perhaps with a comparison table? With that it'd be a better overview I think :)
