### 2020-09-08 #2410 @joannawolthuis
@lpantano Hi! There are still some bugs open, but I know from experience that that can keep going on forever, so I will not hunt for more after the ones I currently addressed are fixed, then that is done :) I would ideally like some improvements to the State of the Field section, I opened an issue here, do you agree @jkanche? https://github.com/girder/viime/issues/624
