### 2020-09-08 #2410 @jkanche
Most of the issues I had before were addressed. I agree with @joannawolthuis on the "Related Work" section. The focus of this section has been based on one tool - MetaboAnalyst. I think it would be great if the authors can talk about other metabolomics tools out there.
