### 2020-09-08 191eb2e
Remove irrelevnat test case
