### 2020-09-10 d176724
PLSDA (#613)
* initial plsda

* add r dependency to base dockerfile

* finish server endpoint

* change endpoint to work with `plotData` mixin

* finish pls-da loadings plot

* get plsda scores plot to display correctly

* fix plsda bug
default number of components for plsda was being computed incorrectly

* allow changing number of components in plsda

* handle num_of_components being too large

* fix lint

* add test for plsda analysis

* fix plsda test and add sample datasets for it

* add some code comments to explain test

* update plsda test to use new fixture

* use append instead of pr
[…]
