+++ viime/analyses.py
@@ -116,3 +116,15 @@ def factor_analysis(measurements: pd.DataFrame, threshold=0.4) -> Dict[str, List
         'threshold': threshold
     })
     return clean(data).to_dict(orient='list')
+
+
+def plsda(measurements: pd.DataFrame, groups: pd.DataFrame, num_of_components: int, mode='scores'):
+    files = {
+        'measurements': measurements.to_csv().encode(),
+        'groups': groups.to_csv(header=True).encode()
+    }
+    data = opencpu_request('plsda', files, {
+        'num_of_components': num_of_components,
+        'mode': mode
+    })
+    return clean(data).to_dict(orient='list')
