### 2020-09-11 c43393f
Add plsda plot default (#666)
