### 2020-09-17 9773308
Add r2 value to response from OPLSDA
