### 2020-09-17 4be6508
Trim NaN value from r2 in OPLSDA
