+++ viime/views.py
@@ -797,7 +797,8 @@ def get_oplsda(validated_table: ValidatedMetaboliteTable):
     column_names = list(measurements.columns)
 
     # TODO: verify this is the correct value to derive "sdev"
-    r2 = modeldf['R2Y']
+    # Last element of r2 is always NaN, slice it off
+    r2 = modeldf['R2Y'][:-1]
     formatted_loadings = []
     x = []
     sdev = []
