+++ web/src/store/index.js
@@ -111,13 +111,6 @@ const plotDefaults = {
     args: {},
     type: plot_types.ANALYSIS,
   },
-  oplsda: {
-    data: null,
-    valid: false,
-    loading: false,
-    args: {},
-    type: plot_types.ANALYSIS,
-  },
   boxplot: {
     data: null,
     valid: false,
