+++ joss/paper.bib
@@ -104,3 +104,93 @@ eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/elps.201500352},
 abstract = {The origin of missing values can be caused by different reasons and depending on these origins missing values should be considered differently and dealt with in different ways. In this research, four methods of imputation have been compared with respect to revealing their effects on the normality and variance of data, on statistical significance and on the approximation of a suitable threshold to accept missing data as truly missing. Additionally, the effects of different strategies for controlling familywise error rate or false discovery and how they work with the different strategies for missing value imputation have been evaluated. Missing values were found to affect normality and variance of data and k-means nearest neighbour imputation was the best method tested for restoring this. Bonferroni correction was the best method for maximizing true positives and minimizing false positives and it was observed that as low as 40\% missing data could be truly missing. The range between 40 and 70\% missing values was defined as a “gray area” and therefore a strategy has been proposed that provides a balance between the optimal imputation strategy that was k-means nearest neighbor and the best approximation of positioning real zeros.},
 year = {2015}
 }
+
+@article{spicer:2017,
+  title={Navigating freely-available software tools for metabolomics analysis},
+  volume={13},
+  DOI={10.1007/s11306-017-1242-7},
+  number={9},
+  journal={Metabolomics},
+  author={Spicer, Rachel and Salek, Reza M. and Moreno, Pablo and Cañueto, Daniel and Steinbeck, Christoph},
+  year={2017},
+}
+
+@article{pluskal:2010,
+  title={MZmine 2: Modular framework for processing, visualizing, and analyzing mass spectrometry-based molecular profile data},
+  volume={11},
+  DOI={10.1186/1471-2105-11-395},
+  number={1},
+  journal={BMC Bioinformatics},
+  author={Pluskal, Tomáš and Castillo, Sandra and Villar-Briones, Alejandro and Orešič, Matej},
+  year={2010},
+}
+
+@article{fernandez:2014,
+  title={An R package to analyse LC/MS metabolomic data: MAIT (Metabolite Automatic Identification Toolkit)},
+  volume={30},
+  DOI={10.1093/bioinformatics/btu136},
+  number={13},
+  journal={Bioinformatics},
+  author={Fernández-Albert, Francesc and Llorach, Rafael and Andrés-Lacueva, Cristina and Perera, Alexandre},
+  year={2014},
+  pages={1937–1939},
+}
+
+@article{melamud:2010,
+  title={Metabolomic Analysis and Visualization Engine for LC−MS Data},
+  volume={82},
+  DOI={10.1021/ac1021166},
+  number={23},
+  journal={Analytical Chemistry},
+  author={Melamud, Eugene and Vastag, Livia and Rabinowitz, Joshua D.},
+  year={2010},
+  pages={9818–9826},
+}
+
+@article{clasquin:2012,
+  title={LC-MS Data Processing with MAVEN: A Metabolomic Analysis and Visualization Engine},
+  DOI={10.1002/0471250953.bi1411s37},
+  journal={Current Protocols in Bioinformatics},
+  author={Clasquin, Michelle F. and Melamud, Eugene and Rabinowitz, Joshua D.},
+  year={2012},
+}
+
+@article{giacomoni:2014,
+  title={Workflow4Metabolomics: a collaborative research infrastructure for computational metabolomics},
+  volume={31},
+  DOI={10.1093/bioinformatics/btu813},
+  number={9}, journal={Bioinformatics},
[…]
