### 2020-09-20 #2410 @waxlamp
@lpantano, @joannawolthuis, @jkanche: we have reworked the Related Work section to discuss a wider range of tools in the space. I think it does a good job of surveying the field and explaining what makes Viime unique.
