### 2020-09-24 7358144
Standardize PLSDA and OPLSDA pages
