### 2020-09-29 fad9ac9
Add group selectors to OPLSDA
