### 2020-09-29 f9d98d9
Return properly formatted VIP scores from PLSDA/OPLSDA
