### 2020-09-30 8ec3efa
Correctly sort default group names for OPLSDA
