### 2020-10-07 #2410 @waxlamp
> @waxlamp, can you create a zenodo archive of the repository and share here? Title and authors should match the paper information. Thanks!

Definitely. The team is working on fixing some typos (no changes to content) and once that's ready, I will take care of this. Thank you!
