### 2020-10-08 37b428b
Use proper parallel construction for numbering steps
Co-authored-by: Brandon Davis <brandon.davis@kitware.com>
