### 2020-10-08 5110b28
Remove stray punctuation
Co-authored-by: Brandon Davis <brandon.davis@kitware.com>
