+++ joss/paper.md
@@ -181,7 +181,7 @@ VIIME supports several downstream analyses and visualizations. Univariate analys
 rank sum test
 and multivariate ANOVA can be carried out on data with two or more groups, respectively.  For
 the ANOVA, a post-hoc Tukey test is automatically applied so that p-values for
-each of the inter-group comparisons are calculated for all metabolites.,
+each of the inter-group comparisons are calculated for all metabolites.
 Metabolites that are significantly different in each of the intergroup
 comparison can be selected for further analysis with a check box at the top of
 each column.  This enables very large datasets with potentially hundreds of
