### 2020-10-08 cd8ae42
Fix grammar
Co-authored-by: Brandon Davis <brandon.davis@kitware.com>
