### 2020-10-09 d3f885c
Expand KNN acronym
