### 2020-10-09 #2410 @waxlamp
> @waxlamp, can you create a zenodo archive of the repository and share here? Title and authors should match the paper information. Thanks!

@lpantano, here you go: https://zenodo.org/record/4075591, with DOI 10.5281/zenodo.4075591.

Please let me know if there's anything else I can do.
