### 2020-10-12 #2410 @lpantano
Thanks, this should work. I just noticed that the version in setup.py is 0.1.0: https://github.com/girder/viime/blob/master/setup.py#L26. Any chance to update that to the right number and update the tar.gz file so the 1.0.0 version match the code version? it is ok if you need to create a 1.0.1 or any other solution. Thanks!
