### 2020-10-13 #2410 @waxlamp
Thank you for catching this, @lpantano. I just reuploaded a fixed version to Zenodo: https://zenodo.org/record/4086165, with DOI 10.5281/zenodo.4086165.

Hopefully this is ok for proceeding. If you think it makes sense, I can also delete the 1.0.0 bundle from Zenodo; please advise.

Thanks!
