### 2020-10-16 #2410 @danielskatz
👋 @waxlamp - There are some small changes needed in the paper.  See https://github.com/girder/viime/pull/687.  Also, in line 194 of that file, is "-log10" correct?  I don't understand the "-".

Finally, for @lpantano - this paper is far, far too long.  Much of the content should be in the documentation, and just pointed to from the paper.  The ideal for a JOSS paper is 250-1000 words, as described in https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain  At this point, we'll let this proceed without asking the authors to cut it, but for next time, please check to rough length of the paper during the review process.
