### 2020-10-18 #2410 @waxlamp
>  @waxlamp - There are some small changes needed in the paper. See [girder/viime#687](https://github.com/girder/viime/pull/687). Also, in line 194 of that file, is "-log10" correct? I don't understand the "-".
>

Thank you!

As @lpantano mentioned, "-log10" refers to negative log-base-10 of the p-value.

@lpantano, now that the requested fixes are applied, what are the next steps?
